Removed the BackgroundColorAnimator, since it makes more since to have the logic inside BottomBar class instead. Also removed the darkTheme / nightMode magic for now.
